Move method to avoid circular dependency
