+++ packages/components/src/internal/components/entities/utils.ts
@@ -15,7 +15,6 @@ import { QueryModel } from '../../../public/QueryModel/QueryModel';
 import { EntityChoice, EntityDataType, IEntityTypeOption } from './models';
 
 import { ParentIdData } from './actions';
-import { Utils } from '@labkey/api';
 
 export function sampleDeleteDependencyText(): string {
     let deleteMsg = '';
@@ -142,21 +141,3 @@ export function getJobCreationHref(
     return actionUrl instanceof AppURL ? actionUrl.toHref() : actionUrl;
 }
 
-export function getPermissionRestrictionMessage(
-    totalCount: number,
-    noPermissionCount: number,
-    nounSingular: string,
-    nounPlural: string,
-    verb: string,
-    verbSuffix?: string
-): string {
-    if (!noPermissionCount) return null;
-
-    const noun = totalCount === 1 ? nounSingular : nounPlural;
-
-    if (noPermissionCount === totalCount)
-        return `You don't have the required permission to ${verb} the selected ${noun}${verbSuffix}.`;
-
-    const notPermittedNoun = Utils.pluralize(noPermissionCount, nounSingular, nounPlural);
-    return `Selection includes ${notPermittedNoun} that you do not have permission to ${verb}${verbSuffix}. Only the ${nounPlural} that you have permission for will be updated.`;
-}
